Rename 'yaml_reader' to 'recipe_reader'
The class 'tool_pluginkenobi_recipe_reader' will do some processing
on the recipe to modify the format to be easier to use in other parts of
the program. For example, after reading the recipe, the 'author' field
will be a numerically-indexed array of associative arrays of exactly one
entry. After processing, 'author' will be changed to an associative array.